Advisor Landing Page - Mostly Implemented
Built most of the backend for the advisor landing page. Created Student and Advisor classes, as well as StudentInformation and Advisor Information. These act as the data layer classes for their corresponding class.

This included adding many SQL changes. These are included in AdvisorLanding_Work.sql in the SQL Scirpts directory.

TODOs:
 - Calculate progress
 - Change progress bar color based on progress completed